Time: 0 ms (100.00%) | Memory: 19.7 MB (43.28%) - LeetSync

+++ 662-maximum-width-of-binary-tree/maximum-width-of-binary-tree.cpp
@@ -37,6 +37,6 @@ public:
             }
                ans= max (ans, last-first+1); 
         }
-        return ans; 
+        return (int)ans; 
     }
 };
